test xong

+++ Ev-System/src/main/java/com/example/Ev/System/repository/PartRepository.java
@@ -4,5 +4,5 @@ import com.example.Ev.System.entity.Part;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PartRepository extends JpaRepository<Part,Integer> {
+public interface PartRepository extends JpaRepository<Part,Long> {
 }
